requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
docstring_parser==0.17.0
elevenlabs==2.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
packaging==25.0
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
